Eliminated BigInt

+++ lab1/src/main/scala/randomized/Primality.scala
@@ -71,7 +71,7 @@ object AKS {
 
     for(a <- 1 until max) {
       val left = Polynomial(Array(a, 1)).pow_mod(n,n) // (x+a)^n // optimize!
-      val right = Polynomial((a +: Seq.fill(n-1)(0) :+ 1).toArray) // (x^n + a)
+      val right = Polynomial((a.toLong +: Seq.fill(n-1)(0L) :+ 1L).toArray) // (x^n + a)
       if(remainder(left, r).mod(n).subtract(remainder(right, r)) != Polynomial(Array(0))) {
         Logger.info("Poly test says: composite!")
         return Some(Composite)
